Basic auth missing from incoming respondent endpoints, and outgoing case requests
