+++ config.ini
@@ -93,7 +93,7 @@ max_upload_file_name_length = 50
 ;
 ;   Values to be shown in /info (the healthcheck endpoint)
 ;
-version = 0.1.0
+version = 0.1.1
 name = collectioninstrumentsvc
 ;
 API_GATEWAY_CASE_URL = http://localhost:8171/cases/
