+++ ras-party-service-preprod_manifest.yml
@@ -1,0 +1,36 @@
+applications:
+- name: ras-party-service-preprod
+  instances: 3
+  memory: 256M
+  disk_quota: 512M
+  routes:
+  - route: ras-party-service-preprod.apps.prod.cf5.onsclofo.uk
+  env:
+    ONS_ENV: preprod
+    api-gateway.host: ras-api-gateway-preprod.apps.prod.cf5.onsclofo.uk
+    api-gateway.port: "80"
+    case-service.host: casesvc-preprod.apps.prod.cf5.onsclofo.uk
+    case-service.port: "80"
+    collectionexercise-service.host: collectionexercisesvc-preprod.apps.prod.cf5.onsclofo.uk
+    collectionexercise-service.port: "80"
+    feature.send_email_to_gov_notify: "True"
+    features.gateway_registration: "true"
+    features.send_email_to_gov_notify: "False"
+    frontstage-service.host: ras-frontstage-preprod.apps.prod.cf5.onsclofo.uk
+    frontstage-service.port: "80"
+    gov-uk-notify-service.gov_notify_api_key: 40c343b3-8fda-4f90-8a0d-5fd062aaa1db
+    gov-uk-notify-service.gov_notify_service_id: ce3674b1-7b08-4377-a6a7-05b5722d4ea5
+    gov-uk-notify-service.gov_notify_template_id: 53c59576-8c3b-4298-99d1-b245ddd28500
+    iac-service.host: iacsvc-preprod.apps.prod.cf5.onsclofo.uk
+    iac-service.port: "80"
+    oauth2-service.host: ras-django-preprod.apps.prod.cf5.onsclofo.uk
+    oauth2-service.port: "80"
+    security_user_name: sdc.preprod
+    security_user_password: damage.sublime.sprawl.purify.apposite.verbose
+    service.debug: "True"
+    survey-service.host: surveysvc-preprod.apps.prod.cf5.onsclofo.uk
+    survey-service.port: "80"
+  services:
+  - ras-party-db
+  stack: cflinuxfs2
+  timeout: 180
