+++ setup.py
@@ -4,7 +4,7 @@ import sys
 from setuptools import setup, find_packages
 
 NAME = "swagger_server"
-VERSION = "1.0.0"
+VERSION = "1.0.1"
 
 # To install the library, run the following
 #
