+++ swagger_server/controllers/ciupload_controller.py
@@ -1,20 +1,10 @@
 from flask import request, jsonify, make_response
-from flask_httpauth import HTTPBasicAuth
 from ons_ras_common import ons_env
 
+from swagger_server.controllers.basic_auth import auth
 from .collectioninstrument import CollectionInstrument
 
 collection_instrument = CollectionInstrument()
-
-auth = HTTPBasicAuth()
-
-
-@auth.get_password
-def get_pw(username):
-    config_username = ons_env.security_user_name
-    config_password = ons_env.security_user_password
-    if username == config_username:
-        return config_password
 
 
 @auth.login_required
